refactor/#27: 코드리뷰 반영
