+++ apis/mySize.ts
@@ -1,12 +1,12 @@
 import { client } from 'apis';
-import { BottomSizeInput, MySizeOutput, TopSizeInput } from 'types/mySize/remote';
+import { BottomSizeInput, MySizeResponse, TopSizeInput } from 'types/mySize/client';
 
 export const postMyTopSize = async (myTopSize: TopSizeInput) => {
-  const { data } = await client.post<MySizeOutput>(`/mySize/topSize`, myTopSize);
+  const { data } = await client.post<MySizeResponse>(`/mySize/topSize`, myTopSize);
   return data;
 };
 
 export const postMyBottomSize = async (myBottomSize: BottomSizeInput) => {
-  const { data } = await client.post<MySizeOutput>(`/mySize/bottomSize`, myBottomSize);
+  const { data } = await client.post<MySizeResponse>(`/mySize/bottomSize`, myBottomSize);
   return data;
 };
